Update README.md
I’m a passionate Sales professional with over 4 years of experience helping customers and resolving their concerns with empathy and efficiency. Proud to have maintained a 95% customer satisfaction rate, I thrive on creating positive experiences and building trust with every interaction. Skilled in handling a wide range of inquiries, troubleshooting issues, and collaborating with teams to ensure smooth processes, I’m always looking for ways to improve and make a difference for customers and the company alike.